try using packaged bluez with Rapsbian

+++ bin/oref0-setup.sh
@@ -630,12 +630,15 @@ if [[ $REPLY =~ ^[Yy]$ ]]; then
         bluetoothdminversion=5.47
         bluetoothdversioncompare=$(awk 'BEGIN{ print "'$bluetoothdversion'"<"'$bluetoothdminversion'" }')
         if [ "$bluetoothdversioncompare" -eq 1 ]; then
-            cd $HOME/src/ && wget -4 https://www.kernel.org/pub/linux/bluetooth/bluez-5.47.tar.gz && tar xvfz bluez-5.47.tar.gz || die "Couldn't download bluez"
-            killall bluetoothd &>/dev/null #Kill current running version if its out of date and we are updating it
-            cd $HOME/src/bluez-5.47 && ./configure --enable-experimental --disable-systemd && \
-            make && sudo make install || die "Couldn't make bluez"
-            killall bluetoothd &>/dev/null #Kill current running version if its out of date and we are updating it
-            sudo cp ./src/bluetoothd /usr/local/bin/ || die "Couldn't install bluez"
+            # use packaged bluez with Rapsbian
+            if ! getent passwd pi > /dev/null; then
+                cd $HOME/src/ && wget -4 https://www.kernel.org/pub/linux/bluetooth/bluez-5.47.tar.gz && tar xvfz bluez-5.47.tar.gz || die "Couldn't download bluez"
+                killall bluetoothd &>/dev/null #Kill current running version if its out of date and we are updating it
+                cd $HOME/src/bluez-5.47 && ./configure --enable-experimental --disable-systemd && \
+                make && sudo make install || die "Couldn't make bluez"
+                killall bluetoothd &>/dev/null #Kill current running version if its out of date and we are updating it
+                sudo cp ./src/bluetoothd /usr/local/bin/ || die "Couldn't install bluez"
+            fi
             oref0-bluetoothup
         else
             echo bluez v ${bluetoothdversion} already installed
